// client/src/entities/record.ts
interface Record {
  __v: number;
  _id: string;
  added: Date;
  country: string;
  end_year: number;
  impact: string;
  insight: string;
  intensity: number;
  likelihood: number;
  pestle: string;
  published: Date;
  region: string;
  relevance: number;
  sector: string;
  source: string;
  start_year: number;
  title: string;
  topic: string;
  url: string;
}
export default Record;